@@ -3,30 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.8.0",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.37",
-    "@material-ui/styles": "^4.7.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.1.2",
-    "i18next": "^19.0.2",
-    "i18next-browser-languagedetector": "^4.0.1",
-    "i18next-xhr-backend": "^3.2.2",
-    "node-sass": "^4.13.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-google-login": "^5.0.7",
-    "react-i18next": "^11.2.7",
-    "react-player": "^1.14.2",
-    "react-redux": "^7.1.3",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.3",
-    "redux": "^4.0.4",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0"
+    "@material-ui/core": "4.12.3",
+    "@material-ui/icons": "4.11.2",
+    "@material-ui/lab": "4.0.0-alpha.60",
+    "@material-ui/styles": "4.11.4",
+    "@testing-library/jest-dom": "5.16.1",
+    "@testing-library/react": "12.1.2",
+    "@testing-library/user-event": "13.5.0",
+    "i18next": "21.6.3",
+    "i18next-browser-languagedetector": "6.1.2",
+    "i18next-xhr-backend": "3.2.2",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-google-login": "5.2.2",
+    "react-i18next": "11.15.1",
+    "react-player": "2.9.0",
+    "react-redux": "7.2.6",
+    "react-router": "6.2.1",
+    "react-router-dom": "6.2.1",
+    "react-scripts": "5.0.0",
+    "redux": "4.1.2",
+    "redux-devtools-extension": "2.13.9",
+    "redux-thunk": "2.4.1",
+    "reselect": "4.1.5",
+    "sass": "^1.45.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -51,20 +51,20 @@
     ]
   },
   "devDependencies": {
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.11.4",
-    "@babel/node": "^7.10.5",
-    "@babel/preset-env": "^7.11.0",
-    "@types/jest": "^25.2.1",
-    "@types/node": "^13.11.0",
-    "@types/react": "^16.9.32",
-    "@types/react-dom": "^16.9.6",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router-dom": "^5.1.3",
-    "coveralls": "^3.1.0",
-    "eslint-plugin-react-hooks": "^2.3.0",
-    "jest-localstorage-mock": "^2.4.0",
-    "redux-logger": "^3.0.6",
-    "typescript": "^3.8.3"
+    "@babel/cli": "7.16.0",
+    "@babel/core": "7.16.5",
+    "@babel/node": "7.16.5",
+    "@babel/preset-env": "7.16.5",
+    "@types/jest": "27.0.3",
+    "@types/node": "17.0.2",
+    "@types/react": "17.0.37",
+    "@types/react-dom": "17.0.11",
+    "@types/react-redux": "7.1.20",
+    "@types/react-router-dom": "5.3.2",
+    "coveralls": "3.1.1",
+    "eslint-plugin-react-hooks": "4.3.0",
+    "jest-localstorage-mock": "2.4.18",
+    "redux-logger": "3.0.6",
+    "typescript": "4.5.4"
   }
 }
