@@ -1,6 +1,7 @@
 import React from "react";
 import { shallowEqual, useSelector } from "react-redux";
-import { useLocation, Redirect, Switch, Route } from "react-router-dom";
+import { useLocation, Navigate, Routes, Route } from "react-router-dom";
+
 import CssBaseline from "@material-ui/core/CssBaseline";
 import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
@@ -45,21 +46,19 @@ const Authorization = () => {
   return (
     <>
       {loggedStatus && (
-        <Redirect
-          to={{
-            pathname: "/",
-            state: { from: location },
-          }}
-        />
+        <Routes>
+          <Route
+            children={() => <Navigate to="/" replace state={{ from: location }}/>}
+          />
+        </Routes>
       )}
       <Grid container component="main" className={classes.root}>
         <CssBaseline />
         <Grid item xs={false} sm={4} md={7} className={classes.image} />
         <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
-          <Switch>
-            <Route exact path="/signup" component={SignUpForm} />
-            <Route exact path="/login" component={LogInForm} />
-          </Switch>
+          <Routes>
+            <Route path="/*" element={location.pathname === '/signup' ? <SignUpForm/> : <LogInForm />} />
+          </Routes>
         </Grid>
       </Grid>
     </>
